--- onedal/backend/data/data.h
@@ -1,34 +1,0 @@
-/*******************************************************************************
-* Copyright 2021 Intel Corporation
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-*     http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*******************************************************************************/
-
-#pragma once
-
-#ifndef ONEDAL_DATA_PARALLEL
-#define NO_IMPORT_ARRAY
-#endif
-
-#include "common/backend/utils.h"
-#include "backend/data/utils.h"
-
-#include <numpy/arrayobject.h>
-#include "oneapi/dal/table/common.hpp"
-
-namespace oneapi::dal::python {
-ONEDAL_BACKEND_EXPORT oneapi::dal::table convert_to_table(PyObject* nda);
-
-ONEDAL_BACKEND_EXPORT PyObject* convert_to_numpy(const oneapi::dal::table& input);
-
-} // namespace oneapi::dal::python
