--- onedal/common/__init__.py
@@ -1,34 +1,0 @@
-#===============================================================================
-# Copyright 2021 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#===============================================================================
-
-from .validation import (
-    _column_or_1d,
-    _validate_targets,
-    _check_X_y,
-    _check_array,
-    _check_is_fitted,
-    _check_classification_targets,
-    _type_of_target,
-    _is_integral_float,
-    _is_multilabel,
-    _check_n_features,
-    _num_features
-)
-
-__all__ = ['_column_or_1d', '_validate_targets', '_check_X_y',
-           '_check_array', '_check_is_fitted',
-           '_check_classification_targets', '_type_of_target', '_is_integral_float',
-           '_is_multilabel', '_check_n_features', '_num_features']
