+++ onedal/datatypes/__init__.py
@@ -14,4 +14,21 @@
 # limitations under the License.
 #===============================================================================
 
-__all__ = ['primitives', 'svm']
+from .validation import (
+    _column_or_1d,
+    _validate_targets,
+    _check_X_y,
+    _check_array,
+    _check_is_fitted,
+    _check_classification_targets,
+    _type_of_target,
+    _is_integral_float,
+    _is_multilabel,
+    _check_n_features,
+    _num_features
+)
+
+__all__ = ['_column_or_1d', '_validate_targets', '_check_X_y',
+           '_check_array', '_check_is_fitted',
+           '_check_classification_targets', '_type_of_target', '_is_integral_float',
+           '_is_multilabel', '_check_n_features', '_num_features']
