+++ onedal/datatypes/data_conversion.hpp
@@ -1,0 +1,29 @@
+/*******************************************************************************
+* Copyright 2021 Intel Corporation
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*******************************************************************************/
+
+#pragma once
+
+#include <Python.h>
+
+#include "oneapi/dal/table/common.hpp"
+
+namespace oneapi::dal::python {
+
+PyObject *convert_to_numpy(const dal::table &input);
+dal::table convert_to_table(PyObject *obj);
+
+} // namespace oneapi::dal::python
+
