+++ onedal/modules.hpp
@@ -1,0 +1,38 @@
+/*******************************************************************************
+* Copyright 2021 Intel Corporation
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*     http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*******************************************************************************/
+
+#pragma once
+
+#include "onedal/common/pybind11_helpers.hpp"
+
+namespace oneapi::dal::python {
+
+/* common */
+ONEDAL_PY_INIT_MODULE(policy);
+
+/* datatypes*/
+ONEDAL_PY_INIT_MODULE(table);
+
+/* primitives */
+ONEDAL_PY_INIT_MODULE(linear_kernel);
+ONEDAL_PY_INIT_MODULE(rbf_kernel);
+ONEDAL_PY_INIT_MODULE(polynomial_kernel);
+ONEDAL_PY_INIT_MODULE(sigmoid_kernel);
+
+/* svm */
+ONEDAL_PY_INIT_MODULE(svm);
+
+} // namespace oneapi::dal::python
