+++ setup.py
@@ -21,7 +21,6 @@ import sys
 import sysconfig
 import time
 from setuptools import setup, Extension
-import setuptools.command.install as orig_install
 import setuptools.command.develop as orig_develop
 import distutils.command.build as orig_build
 from os.path import join as jp
@@ -31,6 +30,7 @@ import glob
 import numpy as np
 import scripts.build_backend as build_backend
 from scripts.version import get_onedal_version
+import pybind11
 
 try:
     from ctypes.utils import find_library
@@ -159,17 +159,21 @@ def get_libs(iface='daal'):
         major_version = get_win_major_version()
         libraries_plat = [f'onedal_core_dll{major_version}']
         onedal_lib = [f'onedal_dll{major_version}']
+        onedal_dpc_lib = [f'onedal_dpc_dll{major_version}']
     else:
         libraries_plat = ['onedal_core', 'onedal_thread']
         onedal_lib = ['onedal']
+        onedal_dpc_lib = ['onedal_dpc']
     if iface == 'onedal':
         libraries_plat = onedal_lib + libraries_plat
+    elif iface == 'onedal_dpc':
+        libraries_plat = onedal_dpc_lib + libraries_plat
     return libraries_plat
 
 
 def get_build_options():
     include_dir_plat = [os.path.abspath('./src'),
-                        os.path.abspath('./onedal'),
+                        os.path.abspath('.'),
                         dal_root + '/include', ]
     # FIXME it is a wrong place for this dependency
     if not no_dist:
@@ -208,54 +212,11 @@ def get_build_options():
     return eca, ela, include_dir_plat
 
 
-def get_sources_onedal():
-    from distutils.dir_util import create_tree
-    from distutils.file_util import copy_file
-
-    cpp_files = glob.glob("onedal/**/**/*.cpp")
-    pyx_files = glob.glob("onedal/**/*.pyx")
-    pxi_files = glob.glob("onedal/**/*.pxi")
-
-    create_tree('build', pyx_files)
-    for f in pyx_files:
-        copy_file(f, jp('build', f))
-
-    main_pyx = 'onedal/onedal.pyx'
-    main_host_pyx = 'build/onedal/onedal_host.pyx'
-    main_dpc_pyx = 'build/onedal/onedal_dpc.pyx'
-    copy_file(main_pyx, main_host_pyx)
-    copy_file(main_pyx, main_dpc_pyx)
-
-    for f in pxi_files:
-        copy_file(f, jp('build', f))
-
-    return cpp_files, main_host_pyx, main_dpc_pyx
-
-
 def getpyexts():
     eca, ela, include_dir_plat = get_build_options()
-    onedal_libraries_plat = get_libs("onedal")
     libraries_plat = get_libs("daal")
-    cpp_files, main_host_pyx, main_dpc_pyx = get_sources_onedal()
 
     exts = []
-
-    ext = Extension('_onedal4py_host',
-                    sources=[main_host_pyx] + cpp_files,
-                    include_dirs=include_dir_plat + [np.get_include()],
-                    extra_compile_args=eca,
-                    extra_link_args=ela,
-                    define_macros=[
-                        ('NPY_NO_DEPRECATED_API',
-                         'NPY_1_7_API_VERSION'),
-                        ('ONEDAL_VERSION', ONEDAL_VERSION),
-                    ],
-                    libraries=onedal_libraries_plat,
-                    library_dirs=ONEDAL_LIBDIRS,
-                    language='c++')
-
-    if is_onedal_iface:
-        exts.extend(cythonize(ext, compile_time_env={'ONEDAL_VERSION': ONEDAL_VERSION}))
 
     ext = Extension('_daal4py',
                     [os.path.abspath('src/daal4py.cpp'),
@@ -271,7 +232,7 @@ def getpyexts():
                     language='c++')
     exts.extend(cythonize(ext))
 
-    if dpcpp and is_onedal_iface:
+    if dpcpp:
         if IS_LIN or IS_MAC:
             runtime_library_dirs = ["$ORIGIN/onedal"]
             runtime_oneapi_dirs = ["$ORIGIN/daal4py/oneapi"]
@@ -279,22 +240,6 @@ def getpyexts():
             runtime_library_dirs = []
             runtime_oneapi_dirs = []
 
-        ext = Extension('_onedal4py_dpc',
-                        sources=[main_dpc_pyx],
-                        include_dirs=include_dir_plat + [np.get_include()],
-                        extra_compile_args=eca,
-                        extra_link_args=ela,
-                        define_macros=[
-                            ('NPY_NO_DEPRECATED_API',
-                             'NPY_1_7_API_VERSION'),
-                        ],
-                        libraries=['dpc_backend'],
-                        library_dirs=['onedal'],
-                        runtime_library_dirs=runtime_library_dirs,
-                        language='c++')
-
-        if is_onedal_iface:
-            exts.extend(cythonize(ext))
         ext = Extension('_oneapi',
                         [os.path.abspath('src/oneapi/oneapi.pyx'), ],
                         depends=['src/oneapi/oneapi.h', 'src/oneapi/oneapi_backend.h'],
@@ -360,55 +305,128 @@ def gen_pyx(odir):
 gen_pyx(os.path.abspath('./build'))
 
 
-def build_oneapi_backend():
+def build_cpp(cc, cxx, sources, targetprefix, targetname, targetsuffix, libs, includes, eca, ela, defines, installpath=''):
     import shutil
     import subprocess
+    from sysconfig import get_paths as gp
+    from os.path import basename
 
-    eca, ela, include_dir_plat = get_build_options()
-    libraries_plat = get_libs('daal')
-    libraries = libraries_plat + ['OpenCL', 'onedal_sycl']
-    include_dir_plat = ['-I' + incdir for incdir in include_dir_plat]
-    library_dir_plat = ['-L' + libdir for libdir in ONEDAL_LIBDIRS]
+    print(f'building cpp target {targetname}...')
+
+    include_dir_plat = ['-I' + incdir for incdir in includes]
+    libdirs = ONEDAL_LIBDIRS
     if IS_WIN:
         eca += ['/EHsc']
-        ela += ['/MD']
         lib_prefix = ''
         lib_suffix = '.lib'
-        libname = 'oneapi_backend.dll'
-        additional_linker_opts = ['/link', '/DLL', f'/OUT:{libname}']
+        obj_ext = '.obj'
+        libdirs += [jp(gp()['data'], 'libs')]
+        library_dir_plat = ['/link'] + [f'/LIBPATH:{libdir}' for libdir in libdirs]
+        additional_linker_opts = ['/DLL', f'/OUT:{targetprefix}{targetname}{targetsuffix}']
     else:
         eca += ['-fPIC']
         ela += ['-shared']
-        lib_suffix = ''
         lib_prefix = '-l'
-        libname = 'liboneapi_backend.so'
-        additional_linker_opts = ['-o', libname]
-    libraries = [f'{lib_prefix}{str(item)}{lib_suffix}' for item in libraries]
+        lib_suffix = ''
+        obj_ext = '.o'
+        library_dir_plat = ['-L' + libdir for libdir in libdirs]
+        additional_linker_opts = ['-o', f'{targetprefix}{targetname}{targetsuffix}']
+    eca += ['-c']
+    libs = [f'{lib_prefix}{str(item)}{lib_suffix}' for item in libs]
 
     d4p_dir = os.getcwd()
-    src_dir = os.path.join(d4p_dir, "src/oneapi")
-    build_dir = os.path.join(d4p_dir, "build_backend")
+    build_dir = os.path.join(d4p_dir, f"build_{targetname}")
 
     if os.path.exists(build_dir):
         shutil.rmtree(build_dir)
     os.mkdir(build_dir)
     os.chdir(build_dir)
 
-    cmd = ['dpcpp'] + \
-        include_dir_plat + eca + \
-        library_dir_plat + ela + \
-        [f'{src_dir}/oneapi_backend.cpp'] + \
-        libraries + additional_linker_opts
+    objfiles = [basename(f).replace('.cpp', obj_ext) for f in sources]
+    for i, cppfile in enumerate(sources):
+        if IS_WIN:
+            out = [f'/Fo{objfiles[i]}']
+        else:
+            out = ['-o', objfiles[i]]
+        cmd = [cc] + include_dir_plat + eca + [f'{d4p_dir}/{cppfile}'] + out + defines
+        print(subprocess.list2cmdline(cmd))
+        subprocess.check_call(cmd)
 
+    cmd = [cxx] + objfiles + library_dir_plat + \
+          ela + libs + additional_linker_opts
     print(subprocess.list2cmdline(cmd))
     subprocess.check_call(cmd)
 
-    shutil.copy(libname, os.path.join(d4p_dir, "daal4py/oneapi"))
+    shutil.copy(f'{targetprefix}{targetname}{targetsuffix}', os.path.join(d4p_dir, installpath))
     if IS_WIN:
-        shutil.copy(libname.replace('.dll', '.lib'),
-                    os.path.join(d4p_dir, "daal4py/oneapi"))
+        target_lib_suffix = targetsuffix.replace('.dll', '.lib')
+        shutil.copy(f'{targetprefix}{targetname}{target_lib_suffix}', os.path.join(d4p_dir, installpath))
     os.chdir(d4p_dir)
 
+
+def build_oneapi_backend():
+    eca, ela, includes = get_build_options()
+
+    return build_cpp(
+        cc='dpcpp',
+        cxx='dpcpp',
+        sources=['src/oneapi/oneapi_backend.cpp'],
+        targetname='oneapi_backend',
+        targetprefix='' if IS_WIN else 'lib',
+        targetsuffix='.dll' if IS_WIN else '.so',
+        libs=get_libs('daal') + ['OpenCL', 'onedal_sycl'],
+        includes=includes,
+        eca=eca,
+        ela=ela,
+        defines=[],
+        installpath='daal4py/oneapi/'
+        )
+
+
+def build_onedal(iface):
+    from sysconfig import get_paths as gp
+    eca, ela, includes = get_build_options()
+    defines = ['-DNPY_NO_DEPRECATED_API=NPY_1_7_API_VERSION',
+               f'-DONEDAL_VERSION={ONEDAL_VERSION}']
+    if iface == 'host':
+        if IS_LIN or IS_MAC:
+            cc, cxx = 'gcc', 'g++'
+            eca += ['-std=c++17', '-O3', '-fvisibility=hidden', '-g0']
+        if IS_MAC:
+            cc, cxx = 'gcc', 'g++'
+            eca += ['-std=c++17', '-O3', '-fvisibility=hidden', '-g0', '-undefined', 'dynamic_lookup']
+            ela += ['-undefined', 'dynamic_lookup']
+        if IS_WIN:
+            cc, cxx = 'cl', 'cl'
+        libs = get_libs('onedal')
+    elif iface == 'dpc':
+        cc, cxx = 'dpcpp', 'dpcpp'
+        defines += ['-DONEDAL_DATA_PARALLEL']
+        libs = get_libs('onedal_dpc')
+
+    build_cpp(
+        cc=cc,
+        cxx=cxx,
+        sources=sorted(glob.glob('onedal/**/*.cpp', recursive=True)),
+        targetname=f'_onedal_py_{iface}',
+        targetprefix='',
+        targetsuffix=get_config_vars('EXT_SUFFIX')[0],
+        libs=libs,
+        includes=includes + [np.get_include()] + [pybind11.get_include()] + [gp()['include']],
+        eca=eca,
+        ela=ela,
+        defines=defines,
+        installpath='onedal'
+        )
+
+
+def get_onedal_py_libs():
+    ext_suffix = get_config_vars('EXT_SUFFIX')[0]
+    libs = [f'_onedal_py_host{ext_suffix}', f'_onedal_py_dpc{ext_suffix}']
+    if IS_WIN:
+        ext_suffix_lib = ext_suffix.replace('.dll', '.lib')
+        libs += [f'_onedal_py_host{ext_suffix_lib}', f'_onedal_py_dpc{ext_suffix_lib}']
+    return libs
 
 def distutils_dir_name(dname):
     """Returns the name of a distutils build directory"""
@@ -418,30 +436,25 @@ def distutils_dir_name(dname):
                     version=sys.version_info)
 
 
-class install(orig_install.install):
-    def run(self):
-        if dpcpp and is_onedal_iface:
-            build_oneapi_backend()
-            if is_onedal_iface:
-                build_backend.custom_build_cmake_clib()
-        return super().run()
-
-
 class develop(orig_develop.develop):
     def run(self):
-        if dpcpp and is_onedal_iface:
+        if is_onedal_iface:
+            build_onedal('host')
+        if dpcpp:
             build_oneapi_backend()
             if is_onedal_iface:
-                build_backend.custom_build_cmake_clib()
+                build_onedal('dpc')
         return super().run()
 
 
 class build(orig_build.build):
     def run(self):
-        if dpcpp and is_onedal_iface:
+        if is_onedal_iface:
+            build_onedal('host')
+        if dpcpp:
             build_oneapi_backend()
             if is_onedal_iface:
-                build_backend.custom_build_cmake_clib()
+                build_onedal('dpc')
         return super().run()
 
 
@@ -475,7 +488,7 @@ setup(
     author_email="scripting@intel.com",
     maintainer_email="onedal.maintainers@intel.com",
     project_urls=project_urls,
-    cmdclass={'install': install, 'develop': develop, 'build': build},
+    cmdclass={'develop': develop, 'build': build},
     classifiers=[
         'Development Status :: 5 - Production/Stable',
         'Environment :: Console',
@@ -520,20 +533,16 @@ setup(
         'daal4py.sklearn.model_selection',
         'onedal',
         'onedal.svm',
-        'onedal.prims',
-        'onedal.common',
+        'onedal.primitives',
+        'onedal.datatypes'
     ],
     package_data={
-        'onedal': [
-            'libdpc_backend.so',
-            'dpc_backend.lib',
-            'dpc_backend.dll'
-        ],
         'daal4py.oneapi': [
             'liboneapi_backend.so',
             'oneapi_backend.lib',
-            'oneapi_backend.dll'
+            'oneapi_backend.dll',
         ],
+        'onedal': get_onedal_py_libs()
     },
     ext_modules=getpyexts()
 )
